restructure the training and serving, also the reload function for neuron_train
