+++ bittensor/_neuron/text/ddp_miner/__init__.py
@@ -101,11 +101,12 @@ class neuron:
         parser.add_argument('--neuron.blacklist', type=float, help='Amount of stake (tao) in order not to get blacklisted', default=0)
         parser.add_argument('--neuron.blacklist_allow_non_registered', action='store_true', help='''If true, black lists non-registered peers''', default=True)
         parser.add_argument('--neuron.sync_block_time', type=int, help='How often the sync the neuron with metagraph, in terms of block time', default=100)
-        parser.add_argument('--neuron.no_restart', action='store_true', help='If True, train the neuron from the beginning', default=False)
+        parser.add_argument('--neuron.no_restart', action='store_false', help='If True, train the neuron from the beginning', default=True)
         parser.add_argument('--neuron.use_wandb', action='store_true', help='''neuron activates its weights and biases powers''', default=False)
         parser.add_argument('--neuron.use_upnpc', action='store_true', help='''neuron attempts to port forward axon using upnpc.''', default=False)
         parser.add_argument('--neuron.multiprocessing', action='store_true', help='Use multiprocessing or not', default=False )
         parser.add_argument('--neuron.world_size', type=int, help='The number of processes for multiprocessing', default=1 )
+        parser.add_argument('--neuron.max_n', type=int, help='The maxinum number of peers in the metagraph to consider', default=4000 )
 
 
         bittensor.logging.add_args( parser )
