+++ bittensor/_neuron/text/ddp_miner/main.py
@@ -1,3 +1,3 @@
 import bittensor
 if __name__ == "__main__":
-    template = bittensor.neurons.ddp_miner.neuron().run_parallel()
+    template = bittensor.neurons.ddp_miner.neuron().run()
