+++ bittensor/_neuron/text/ddp_miner/neuron_impl.py
@@ -22,15 +22,12 @@ Example:
     $ neurons.text.base_neuron_v1().run()
 
 """
-import os
+
 import pandas
 from pandas.core.frame import DataFrame
 import bittensor
 import math
 import torch
-import torch.multiprocessing as mp
-from torch.nn.parallel import DistributedDataParallel as DDP
-import torch.distributed as dist
 import traceback
 import sys
 import wandb
@@ -47,584 +44,19 @@ import torch.nn as nn
 from functools import partial
 
 import torch.nn.functional as F
-from torch.multiprocessing import Manager
-
-
+from .neuron_serve_impl import NeuronServe
+from .neuron_train_impl import DDPNeuronTrain
+from multiprocessing import Process
 class Neuron:
-
     def __init__( self, config: 'bittensor.config', nucleus: 'Nucleus'):
-        r""" Initializes the neuron with the passed config.
-        """
         self.config = config
-        self.world_size = config.neuron.world_size# torch.cuda.device_count()
         self.wallet = bittensor.wallet ( config = self.config )
-        self.nucleus_base = nucleus
-        self.stats = SimpleNamespace(
-            global_step = 0,
-            last_sync_block = 0,
-            epoch_data_size = 0,
-            epoch_sync_count = 0,
-            local_target_epoch_loss = math.inf,
-            distillation_epoch_loss = math.inf,
-            remote_target_epoch_loss = math.inf,
-            local_epoch_acc = 0,
-            best_epoch_loss = math.inf,
-            scores = {},
-            ema_scores = torch.nn.Parameter(torch.zeros(0), requires_grad = False)
-        )
-        # # ---- Decay factor for fisher ema score 
-        self.fisher_ema_decay = 0.995
-
-    def __enter__(self):
+        self.server = NeuronServe(config, self.wallet)
+        self.trainer = DDPNeuronTrain(config, nucleus, self.wallet)
         self.wallet.create()
 
-    def __exit__ ( self, exc_type, exc_value, exc_traceback ):
-        del self.dendrite
-        self.dataset.close()
-        self.axon.stop()   
-        print(exc_type, exc_value, exc_traceback)
-    
-    def init_process(self, rank):
-        os.environ['MASTER_ADDR'] = 'localhost'
-        os.environ['MASTER_PORT'] = '8865'
-        if 'cuda' in self.config.neuron.device:
-            backend = 'nccl'
-        else:
-            backend = 'gloo'
-
-        dist.init_process_group(backend, rank=rank, world_size=self.world_size)
-    
-    def cleanup(self):
-        dist.destroy_process_group()
-
-    def run_parallel( self ):
-        with Manager() as manager:
-            self.stats.scores = manager.dict()
-            mp.spawn(self.run,
-                args=(self.world_size,),
-                nprocs=self.world_size,
-                join=True
-            )
-
-    def init_bit(self, rank = 0):
-        if self.config.neuron.multiprocessing and self.config.neuron.device == 'cuda':
-            self.device = torch.device( device = f'cuda:{rank}' )
-        else:
-            self.device = torch.device( device = self.config.neuron.device )
-        
-        self.subtensor = bittensor.subtensor ( config = self.config )
-        self.metagraph = bittensor.metagraph ( config = self.config, subtensor = self.subtensor )
-        self.dendrite = bittensor.dendrite ( config = self.config, wallet = self.wallet )
-        self.dataset = bittensor.dataset ( config = self.config, world_size = self.world_size, rank = rank)
-        # self.axon = bittensor.axon (
-        #     config = self.config,
-        #     wallet = self.wallet,
-        #     forward_text = self.forward_text,
-        #     backward_text = self.backward_text,
-        #     blacklist = self.blacklist,
-        # )
-        self.stats.ema_scores.to(self.device)
-        
-        if rank == 0 :
-            self.subtensor.register( self.wallet )
-            # self.axon.start().serve (
-            #     use_upnpc = self.config.neuron.use_upnpc, 
-            #     subtensor = self.subtensor
-            # )
-    def clip_grad_hook( self, process_group: dist.ProcessGroup, bucket: dist.GradBucket ): # -> torch.futures.Future[torch.Tensor]:
-        print("in all reduce bucket index", bucket.index())
-        
-        if len(bucket.parameters()) > 70:
-            total_norm = clip_grad_norm_(bucket.parameters(), 1)
-            self.total_norm = total_norm
-            print("total norm", total_norm, bucket.index(), len(bucket.parameters()))
-        else:
-            clip_coef = 1 / (self.total_norm + 1e-6)
-            clip_coef_clamped = torch.clamp(clip_coef, max=1.0)
-            for p in bucket.parameters():
-                p.grad.detach().mul_(clip_coef_clamped)
-            print("reusing norm", self.total_norm, bucket.index(),len(bucket.parameters()))
-            
-        flat_grads = [ torch.reshape(p.grad, (-1,) ) for p in bucket.parameters()]
-        tensor = torch.cat(flat_grads)
-        bucket.set_buffer(tensor)
-        group_to_use = process_group if process_group is not None else dist.group.WORLD
-
-        # Apply the division first to avoid overflow, especially for FP16.
-        tensor.div_(group_to_use.size())
-
-        return (
-            dist.all_reduce(tensor, group=group_to_use, async_op=True)
-            .get_future()
-            .then(lambda fut: fut.value()[0])
-        )
-
-    def run( self, rank = 0, world_size = 0):
-        r""" Miner main loop.
-        """
-
-        # ---- Build Bittensor neuron ----
-        with self:
-            self.init_bit(rank)
-            bittensor.logging.success( prefix = f'Finished initialization', sufix = f'Rank {rank}')
-            if self.config.neuron.use_wandb:
-                bittensor.wandb(
-                    config = self.config,
-                    cold_pubkey = self.wallet.coldkeypub.ss58_address,
-                    hot_pubkey = self.wallet.hotkey.ss58_address,
-                    root_dir = self.config.neuron.full_path
-                )
-
-            # ---- Init run state ----
-            self.epoch = 0   
-
-            # ---- reloads previous run if not restart ----
-            if self.config.neuron.no_restart:
-                self.save()
-
-            try:
-                self.reload(rank)
-            except Exception as e:
-                logger.error("Error when trying to reload model: {}".format(e))
-                self.save()
-                self.reload(rank)
-        
-            if self.config.neuron.multiprocessing:
-                self.init_process(rank)
-                bittensor.logging.success( prefix = 'Inited process group', sufix = '<blue>Rank: {}</blue>'.format( rank ))
-                if 'cuda' in self.config.neuron.device:
-                    self.nucleus.to(rank)
-                    self.nucleus = DDP(self.nucleus,  device_ids=[rank])
-                else:
-                    self.nucleus = DDP(self.nucleus, bucket_cap_mb = 10000000)
-                    self.nucleus.register_comm_hook(state=None, hook=self.clip_grad_hook)
-                    
-                
-                bittensor.logging.success( prefix = 'Enabled DDP', sufix = '<blue>Rank: {}</blue>'.format( rank ))
-
-            self.nucleus.metagraph = self.metagraph
-            self.nucleus.dendrite = self.dendrite
-            bittensor.logging.success( prefix = f'finished setting ddp dend and meta', sufix = f'Rank {rank}')
-
-            optimizer = torch.optim.SGD(
-                # [ {'params': nucleus_ddp.peer_weights, 'lr': self.config.neuron.learning_rate_chain} ],
-                [ {'params': self.nucleus.parameters(), 'lr': self.config.neuron.learning_rate_chain} ],
-                lr = self.config.neuron.learning_rate,
-                momentum = self.config.neuron.momentum,
-            )
-
-            bittensor.logging.success( prefix = f'Initialized', sufix = f'Rank {rank}')
-            
-            self.stats.ema_scores = torch.nn.Parameter(torch.ones(self.metagraph.n.item()).to(self.device) * (1 / self.metagraph.n.item()), requires_grad = False)
-
-            # --- Run until n_epochs ----
-            while self.epoch < self.config.neuron.n_epochs:
-                try:
-                    # --- Init epoch stat----
-                    self.stats.epoch_data_size = 0
-                    self.stats.epoch_sync_count = 0
-                    total_local_target_epoch_loss = 0
-                    total_distillation_epoch_loss = 0
-                    total_remote_target_epoch_loss = 0
-                    total_local_epoch_acc = 0
-                    batches_count = 0
-                    bittensor.logging.success( prefix = f'reseting stats to zeros', sufix = f'batches_count {batches_count}, Rank {rank}')
-
-                    # ---- Run epoch ----
-                    start_block = self.subtensor.get_current_block() + 1
-                    end_block = start_block + self.config.neuron.epoch_length
-                    block_steps = [ block_delta for block_delta in range(start_block, end_block)]
-                    progress_bar = qqdm( block_steps, total=len(block_steps), desc=format_str('blue', f'Epoch:'))
-                    progress_bar.set_bar = partial(progress_bar.set_bar,  element='#')
-                    for block in progress_bar:
-                        # --- Iterate over batches until the end of the block.
-                        current_block = self.subtensor.get_current_block()
-                        while block >= current_block:
-                            # ---- Forward pass ----
-                            inputs = next( self.dataset )
-                            output = self.nucleus.forward(
-                                inputs = inputs.to( self.device ),
-                                training = True,
-                            )
-                            bittensor.logging.success( prefix = f'Forward pass', sufix = f'batches_count {batches_count}, Rank {rank}')
-
-                            # ---- Backward pass ----
-                            output.loss = output.local_target_loss + output.distillation_loss + output.remote_target_loss
-                            output.loss.backward(retain_graph = True) # Accumulates gradients on the nucleus.
-                            bittensor.logging.success( prefix = f'Backward pass', sufix = f'batches_count {batches_count}, Rank {rank}')
-                            # clip_grad_norm_(self.nucleus.parameters(), self.config.neuron.clip_gradients)
-                            
-                            # ---- Apply and zero accumulated gradients.
-                            optimizer.step() 
-                            optimizer.zero_grad()
-                            bittensor.logging.success( prefix = f'Optimizer pass', sufix = f'batches_count {batches_count}, Rank {rank}')
-
-                            current_block = self.subtensor.get_current_block()
-
-                            # ---- Aggrigate outputs and losses 
-                            total_local_target_epoch_loss += output.local_target_loss.item()
-                            total_distillation_epoch_loss += output.distillation_loss.item()
-                            total_remote_target_epoch_loss += output.remote_target_loss.item()
-                            total_local_epoch_acc += output.local_accuracy
-                            self.stats.epoch_data_size += inputs.nelement()
-                            batches_count += 1
-                            
-                            bittensor.logging.success( prefix = f'adding to local target epoch loss', sufix = f' {output.local_target_loss.item(), total_local_target_epoch_loss}, Rank {rank}')
-                            bittensor.logging.success( prefix = f'adding to distillation epoch loss', sufix = f' {output.distillation_loss.item(), total_distillation_epoch_loss}, Rank {rank}')
-                            bittensor.logging.success( prefix = f'adding to remote target epoch loss', sufix = f' {output.remote_target_loss.item(), total_remote_target_epoch_loss}, Rank {rank}')
-                            bittensor.logging.success( prefix = f'adding to local accuracy', sufix = f' {output.local_accuracy, total_local_epoch_acc}, Rank {rank}')
-                            
-                            # ---- Expand ema_scores tensor if the chain grew and aggrigate the score
-                            output.scores = output.scores.detach()
-                            output.scores.requires_grad = False
-                            self.stats.scores[rank] = output.scores
-                            print(rank, output.scores, self.stats.scores)
-                                              
-                            chain_growth = max(output.scores.shape[0] - self.stats.ema_scores.shape[0], 0)
-                            if chain_growth > 0:
-                                self.stats.ema_scores = torch.nn.Parameter(torch.cat( [self.stats.ema_scores, torch.zeros([chain_growth], dtype=torch.float32, device = self.device)]), requires_grad=False)
-                            
-                            scores_avg =  sum(self.stats.scores.values())/len(self.stats.scores)
-                            print("avg", scores_avg[:20])
-                            self.stats.ema_scores = self.fisher_ema_decay * self.stats.ema_scores + (1 - self.fisher_ema_decay) * scores_avg
-                            
-                        # ---- Update the epoch loss if iVt is the last iteration within epoch
-                        if block+1 == end_block :
-                            self.stats.local_target_epoch_loss = total_local_target_epoch_loss / (batches_count) 
-                            self.stats.distillation_epoch_loss = total_distillation_epoch_loss / (batches_count)
-                            self.stats.remote_target_epoch_loss = total_remote_target_epoch_loss / (batches_count)
-                            self.stats.local_epoch_acc = total_local_epoch_acc / (batches_count)
-                            bittensor.logging.success( prefix = f'getting avg of local target epoch loss', sufix = f' {self.stats.local_target_epoch_loss, total_local_target_epoch_loss, batches_count, self.world_size}, Rank {rank}')
-                            bittensor.logging.success( prefix = f'getting avg of distillation epoch loss', sufix = f' {self.stats.distillation_epoch_loss, total_distillation_epoch_loss, batches_count, self.world_size}, Rank {rank}')
-                            bittensor.logging.success( prefix = f'getting avg of remote_target epoch loss', sufix = f' {self.stats.remote_target_epoch_loss, total_remote_target_epoch_loss, batches_count, self.world_size}, Rank {rank}')
-                            bittensor.logging.success( prefix = f'getting avg of local epoch acc', sufix = f' {self.stats.local_epoch_acc, total_local_epoch_acc, batches_count, self.world_size}, Rank {rank}')
-
-                        if rank == 0:
-                            # ---- Sync with metagraph if the current block >= last synced block + sync block time 
-                            current_block = self.subtensor.get_current_block()
-                            block_diff = current_block - self.stats.last_sync_block
-                            if block_diff >= self.config.neuron.sync_block_time:
-                                self.sync(current_block)                                                                                                                
-                                self.stats.last_sync_block = current_block
-                                self.stats.epoch_sync_count += 1
-                                
-
-                            # ---- Block logs.
-
-                            bittensor.logging.success( prefix = f'local target epoch loss', sufix = f' {self.stats.local_target_epoch_loss}, Rank {rank}')
-                            bittensor.logging.success( prefix = f'distillation epoch loss', sufix = f' {self.stats.distillation_epoch_loss}, Rank {rank}')
-                            bittensor.logging.success( prefix = f'remote target epoch loss', sufix = f' {self.stats.remote_target_epoch_loss}, Rank {rank}')
-                            bittensor.logging.success( prefix = f'local epoch accuracy', sufix = f' {self.stats.local_epoch_acc}, Rank {rank}')
-                            bittensor.logging.success( prefix = f'', sufix = f' Rank {rank}')
-                            bittensor.logging.success( prefix = f'', sufix = f' Rank {rank}')
-                            bittensor.logging.success( prefix = f'', sufix = f' Rank {rank}')
-                            bittensor.logging.success( prefix = f'', sufix = f' Rank {rank}')
-                            bittensor.logging.success( prefix = f'', sufix = f' Rank {rank}')
-                            self.logs (
-                                progress_bar,
-                                iteration = block-start_block,
-                                output = output,
-                            )
-                            self.stats.global_step += 1
-
-                    if rank == 0:
-                        # ---- Checkpoint state ----
-                        self.checkpoint()
-
-                    # ---- Update params ----
-                    self.epoch += 1
-
-                except KeyboardInterrupt:
-                    # --- User ended session ----
-                    self.cleanup()
-                    break
-
-                except Exception as e:
-                    # --- Unknown error ----
-                    logger.exception('Unknown exception: {} with traceback {}', e, traceback.format_exc())
-                    if self.config.neuron.restart_on_failure == True:
-                        logger.info('Restarting from last saved state.')
-                        self.reload()
-                    else:
-                        break
-
-    # ---- Axon Forward call ----
-    def forward_text ( self, inputs_x: torch.FloatTensor) -> torch.FloatTensor:
-        r""" Subscribed to an axon servicing endpoint: processes forward messages from the wire.
-            The arguments reflect an RPC request from another miner in the network, the response tensor
-            should be the hidden units computed using the local context and with shape: [batch_size, sequence_len, __network_dim__].
-
-            Args:
-                inputs_x ( :obj:`torch.Tensor`, `required`):
-                    torch inputs to be forward processed.
-
-            Returns:
-                outputs (:obj:`torch.FloatTensor`):
-                    The nucleus's outputs as a torch tensor of shape [batch_size, sequence_len, __network_dim__]
-        """
-        output = self.nucleus.forward (
-            forward_type = 'local',
-            inputs = inputs_x.to( self.device )
-        )
-        return output.local_hidden
-
-    # ---- Axon Backward call ----
-    def backward_text ( self, inputs_x:torch.FloatTensor, grads_dy:torch.FloatTensor ):
-        r""" Subscribed to an axon servicing endpoint: Processes backward messages from the wire.
-            Arguments reflect an RPC backward request from another miner in the network. No response
-            needed for tokenized text inputs (uint64s have no gradient).
-
-            Args:
-                inputs_x ( :obj:`torch.Tensor`, `required`):
-                    torch inputs from previous forward call.
-                grads_dy ( :obj:`torch.Tensor`, `required`):
-                    torch grads of forward output.                    
-        """
-        if self.config.neuron.accumulate_remote_gradients:
-            with torch.enable_grad():
-                # ---- Set up inputs for gradient computations.
-                outputs_y = self.nucleus.forward(
-                    forward_type = 'local',
-                    inputs = inputs_x.to( self.device ) ).local_context.to( self.device )
-                # ---- The backward call will accumulate gradients on our parameters.
-                torch.autograd.backward (
-                    tensors = [outputs_y],
-                    grad_tensors = [grads_dy.to( self.device )]
-                )
-    
-    def priority(self, pubkey:str, request_type:bittensor.proto.RequestType, inputs_x: torch.FloatTensor) -> float:
-        r"""Return the request priority based on stake and size of input. 
-            Used by the Axon to order requests.
-            Args:
-                pubkey ( str, `required`):
-                    The public ss58 address of the caller.
-                inputs_x ( :obj:`torch.Tensor`, `required`):
-                    torch inputs to be forward processed.
-                request_type ( bittensor.proto.RequestType, `required`):
-                    the request type ('FORWARD' or 'BACKWARD').
-        """        
-        # Priority = stake / request_size 
-        priority = self.metagraph.S[ self.metagraph.hotkeys.index(pubkey) ] / sys.getsizeof(inputs_x)
-        return priority
-
-    def blacklist(self, pubkey:str, request_type:bittensor.proto.RequestType) -> bool:
-        r"""Axon security blacklisting, used to blacklist message from low stake members
-            Currently, this is not turned on.
-            Args:
-                pubkey ( str, `required`):
-                    The public key of the caller.
-                request_type ( bittensor.proto.RequestType, `required`):
-                    the request type ('FORWARD' or 'BACKWARD').
-        """
-        # Blacklist requests from peers who are not subscribed or have stake less that black_list
-        is_registered = pubkey in self.metagraph.hotkeys
-
-        # If we allow non-registered requests return False = not blacklisted.
-        if not is_registered:
-            if self.config.neuron.blacklist_allow_non_registered:
-                return False
-            else:
-                return True
-        else:
-            # Else, get stake and check is above blacklist stake min.
-            uid = self.metagraph.hotkeys.index( pubkey )
-            if self.metagraph.S[uid].item() >= self.config.neuron.blacklist:
-                return False
-            else:
-                return True
-
-    def checkpoint( self ):
-        r""" Optionally Saves, updates and then reloads the miner training state.
-        """
-        last_saved = self.get_saved_state()
-        if last_saved == None or last_saved['epoch_loss'] >= self.stats.local_target_epoch_loss:
-            self.stats.best_epoch_loss = self.stats.local_target_epoch_loss
-            self.save()
-
-        # Checks if epochs managed to diverage
-        if not math.isfinite(self.stats.local_target_epoch_loss):
-            logger.error('Incorrect epoch loss detected, reloading to previous saved state')
-            self.reload()
-
-    def get_saved_state( self ):
-        r""" Returns a saved state dict or none.
-        """
-        try:
-            return torch.load("{}/model.torch".format( self.config.neuron.full_path ))
-        except Exception as e:
-            logger.warning('No saved model found with error: {}', e)
-            logger.info('Initalizing with new model')
-            return None
-
-    def reload( self, rank = 0 ):
-        r""" Reloads/updates the training state from the disk.
-        """
-        # --- Load prev state.
-        state_dict = self.get_saved_state()
-        self.nucleus = self.nucleus_base.to(self.device)
-        
-        # --- Loads and syncs metagraph.
-        try:
-            self.metagraph.sync().save()
-            self.stats.last_sync_block= self.subtensor.get_current_block()
-        except Exception as e:
-            logger.error('Error in loading metagraph: {}'.format(e))
-            self.metagraph.sync().save()
-
-        # ---- Load training state.
-        self.epoch = state_dict['epoch']
-        self.stats.local_target_epoch_loss = state_dict['epoch_loss']
-        self.stats.best_epoch_loss = state_dict['epoch_loss']
-        self.stats.global_step = state_dict['global_step']
-
-        try:
-            self.nucleus.load_state_dict( state_dict['nucleus_state'], strict=False )
-        except Exception as e:
-            logger.exception('Failed to load nucleus state with error, updating the current state')
-            state_dict['nucleus_state'] = self.nucleus.state_dict()
-            torch.save( state_dict, "{}/model.torch".format( self.config.neuron.full_path ) )
-        
-        self.nucleus.to( self.device ) # Load nucleus
-        self.nucleus.device = self.device 
-        self.nucleus.dendrite = self.dendrite # Set local dendrite.
-        self.nucleus.metagraph = self.metagraph_callback # Set local metagraph.
-        if 'optimizer_state' in state_dict.keys():
-            self.optimizer = torch.optim.SGD(
-                [{"params": self.nucleus.parameters()}],
-                lr = state_dict['optimizer_state']['param_groups'][0]['lr'],
-                momentum = state_dict['optimizer_state']['param_groups'][0]['momentum'],
-            )
-        else:
-            self.optimizer = torch.optim.SGD(
-                [{"params": self.nucleus.parameters()}],
-                lr = self.config.neuron.learning_rate,
-                momentum = self.config.neuron.momentum,
-            )
-
-
-        bittensor.logging.success( prefix = 'Reloaded model', sufix = '<blue>{}/model.torch</blue>'.format( self.config.neuron.full_path ))
-
-    def sync (self, current_block ):
-        """ Miner sync with metagraph and update chain weight
-        """
-        # ---- Set weights on chain ----
-        bittensor.logging.success( 'Start syncing metagraph:', 'Block: {}'.format(current_block))
-        self.set_peer_weights()
-
-        # ---- Sync with metagraph ----
-        self.metagraph.sync().save()
-        chain_growth = max(self.metagraph.n.item()- self.stats.ema_scores.shape[0], 0)
-        self.stats.ema_scores = torch.nn.Parameter(torch.cat( [self.stats.ema_scores, torch.zeros([chain_growth], dtype=torch.float32, requires_grad=False).to(self.device)]))
-        bittensor.logging.success( 'Synced metagraph:', 'Block: {}'.format(current_block))
-
-    def save( self ):
-        r""" Saves the training state to disk.
-        """
-        try:
-            state_dict = {
-                'epoch': self.epoch,
-                'epoch_loss': self.stats.local_target_epoch_loss,
-                'global_step': self.stats.global_step,
-                'nucleus_state': self.nucleus.state_dict(), # Save nucleus state.
-                'network': self.subtensor.network # Save Network
-            }
-            if hasattr(self, 'optimizer'):
-                state_dict['optimizer_state']= self.optimizer.state_dict() # Save optimizer.
-            torch.save( state_dict, "{}/model.torch".format( self.config.neuron.full_path ) )
-            bittensor.logging.success(prefix='Saved model', sufix='<blue>{}/model.torch</blue>'.format( self.config.neuron.full_path ) )
-        except Exception as e:
-            logger.exception('Failed to save model with error:{}', e)
-
-    def set_peer_weights( self ):
-        r""" Sets the fisher ema score to peers.
-        """
-
-        try:
-            k = min( self.config.neuron.n_topk_peer_weights, self.metagraph.n.item() )
-            inactive_uids = torch.where(self.metagraph.active == 0)[0]
-            self.stats.ema_scores[inactive_uids] = 0
-            topk_scores, topk_uids = bittensor.unbiased_topk( self.stats.ema_scores , k = k )
-            topk_uids = topk_uids.detach().to('cpu')
-            topk_scores = topk_scores.detach().to('cpu')
-            self.subtensor.set_weights(
-                uids = topk_uids,
-                weights = topk_scores,
-                wait_for_inclusion = False,
-                wallet = self.wallet,
-            )
-
-        except Exception as e:
-            logger.error('Failure setting weights on chain with error: {}', e)
-
-    def metagraph_callback(self):
-        return self.metagraph
-
-    # ---- Training logs ----
-    def logs( self, progress_bar, iteration:int, output: SimpleNamespace ):
-        r""" Called after every training step. Displays miner state to screen.
-        """
-        self_neuron = self.subtensor.neuron_for_pubkey( self.wallet.hotkey.ss58_address )
-        self_uid = self_neuron.uid
-        stake = self_neuron.stake
-        rank = self_neuron.rank
-        incentive = self_neuron.incentive
-        # normalized_peer_weights = F.softmax (self.nucleus.state_dict()['module.peer_weights'], dim=0)
-        current_block = self.subtensor.get_current_block()
-
-        # ---- Progress bar log
-        info = {
-            'Step': colored('{}'.format(self.stats.global_step), 'red'),
-            'Epoch': colored('{}'.format(self.epoch+1), 'yellow'),
-            'Best-loss': colored('{:.4f}'.format(self.stats.best_epoch_loss), 'green'),          
-            'L-loss': colored('{:.4f}'.format(output.local_target_loss.item()), 'blue'),
-            'R-loss': colored('{:.4f}'.format(output.remote_target_loss.item()), 'red'),
-            'D-loss': colored('{:.4f}'.format(output.distillation_loss.item()), 'yellow'),
-            'L-acc': colored('{:.4f}'.format(output.local_accuracy), 'green'),
-            'nPeers': colored(self.metagraph.n.item(), 'blue'),
-            'Stake(\u03C4)': colored('{:.3f}'.format(stake), 'red'),
-            'Rank(\u03C4)': colored('{:.3f}'.format(rank), 'yellow'),
-            'Incentive(\u03C4/block)': colored('{:.6f}'.format(incentive), 'green'),
-            'Current Block': colored('{}'.format(current_block), 'blue'),
-            'Synced Block': colored('{}'.format(self.stats.last_sync_block), 'yellow'),
-        }
-        # ---- Miner summary per peer for progress bar
-        k = min( self.config.neuron.n_topk_peer_weights, self.metagraph.n.item() )
-        topk_scores, topk_uids = bittensor.unbiased_topk( self.stats.ema_scores, k, dim=0 )
-
-        progress_bar.set_infos( info )
-
-        # ---- wandb log if it is the end of epoch 
-        if self.config.neuron.use_wandb and ((iteration + 1) % (self.config.neuron.epoch_length ) == 0):
-            # ---- Miner summary for wandb
-            wandb_info = {
-                'neuron/stake':stake,
-                'neuron/rank':rank,
-                'neuron/incentive':incentive,
-                'neuron/num_peers':self.metagraph.n.item(),
-                'nucleus/remote_target_epoch_loss': self.stats.remote_target_epoch_loss,
-                'nucleus/distillation_epoch_loss': self.stats.distillation_epoch_loss,
-                'nucleus/local_target_epoch_loss': self.stats.local_target_epoch_loss,
-                'nucleus/local_epoch_acc': self.stats.local_epoch_acc,
-                'neuron/num_sync_metagraph': self.stats.epoch_sync_count,
-                'neuron/data_size': self.stats.epoch_data_size,
-            }
-
-            # Build stats dataframe.
-            df = pandas.concat( [
-                bittensor.utils.indexed_values_to_dataframe( prefix = 'fisher_ema_score', index = topk_uids, values = self.stats.ema_scores, filter_zeros = True),
-                # bittensor.utils.indexed_values_to_dataframe( prefix = 'raw_peer_weight', index = topk_uids, values = self.nucleus.state_dict()['module.peer_weights'], filter_zeros = True),
-                # bittensor.utils.indexed_values_to_dataframe( prefix = 'normalized_peer_weight', index = topk_uids, values = normalized_peer_weights, filter_zeros = True),
-                bittensor.utils.indexed_values_to_dataframe( prefix = 'w_{}_i'.format(self_uid), index = topk_uids, values = self.metagraph.W[ self_uid, : ], filter_zeros = True),
-                bittensor.utils.indexed_values_to_dataframe( prefix = 'w_i_{}'.format(self_uid), index = topk_uids, values = self.metagraph.W[ :, self_uid ], filter_zeros = True),
-                # self.axon.to_dataframe( metagraph = self.metagraph ),
-                self.dendrite.to_dataframe( metagraph = self.metagraph )
-            ], axis = 1)
-            df['uid'] = df.index
-            stats_data_table = wandb.Table( dataframe = df)
-
-            # wandb_info_axon = self.axon.to_wandb()
-            wandb_info_dend = self.dendrite.to_wandb()
-            wandb.log( { **wandb_info, **wandb_info_dend }, step = current_block)
-            wandb.log( { 'stats': stats_data_table}, step = current_block)
-            wandb.log( { 'axon_query_times': wandb.plot.scatter( stats_data_table, "uid", "axon_query_time", title="Axon Query time vs UID") } )
-            wandb.log( { 'dendrite_query_times': wandb.plot.scatter( stats_data_table, "uid", "dendrite_query_time", title="Dendrite Query time vs UID") } )
+    def run(self):
+        # p1 = Process(target=self.server.run)
+        p2 = Process(target=self.trainer.run_parallel)
+        # p1.start()
+        p2.start()
